# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/productor/{productor_id}/farms": {
            "get": {
                "tags": [
                    "productor"
                ],
                "summary": "Get Productor Farms",
                "operationId": "get_productor_farms_api_v1_productor__productor_id__farms_get",
                "parameters": [
                    {
                        "name": "productor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Productor Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Farm"
                                    },
                                    "title": "Response Get Productor Farms Api V1 Productor  Productor Id  Farms Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farm/": {
            "get": {
                "tags": [
                    "farm"
                ],
                "summary": "List Farms",
                "operationId": "list_farms_api_v1_farm__get",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Farm"
                                    },
                                    "title": "Response List Farms Api V1 Farm  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "farm"
                ],
                "summary": "Create Farm",
                "operationId": "create_farm_api_v1_farm__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FarmCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farm"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farm/{farm_id}": {
            "get": {
                "tags": [
                    "farm"
                ],
                "summary": "Get Farm",
                "operationId": "get_farm_api_v1_farm__farm_id__get",
                "parameters": [
                    {
                        "name": "farm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Farm Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farm"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "farm"
                ],
                "summary": "Update Farm",
                "operationId": "update_farm_api_v1_farm__farm_id__put",
                "parameters": [
                    {
                        "name": "farm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Farm Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FarmCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farm"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "farm"
                ],
                "summary": "Delete Farm",
                "operationId": "delete_farm_api_v1_farm__farm_id__delete",
                "parameters": [
                    {
                        "name": "farm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Farm Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Farm"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farm/{farm_id}/add-plantation": {
            "put": {
                "tags": [
                    "farm"
                ],
                "summary": "Add Plantation",
                "operationId": "add_plantation_api_v1_farm__farm_id__add_plantation_put",
                "parameters": [
                    {
                        "name": "farm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Farm Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/FarmPlantationSeasonCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FarmPlantationSeason"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farm/{farm_id}/plantations": {
            "get": {
                "tags": [
                    "farm"
                ],
                "summary": "Get Farm Plantations",
                "operationId": "get_farm_plantations_api_v1_farm__farm_id__plantations_get",
                "parameters": [
                    {
                        "name": "farm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Farm Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FarmPlantation"
                                    },
                                    "title": "Response Get Farm Plantations Api V1 Farm  Farm Id  Plantations Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/farm/{farm_id}/seasons": {
            "get": {
                "tags": [
                    "farm"
                ],
                "summary": "Get Farm Seasons",
                "operationId": "get_farm_seasons_api_v1_farm__farm_id__seasons_get",
                "parameters": [
                    {
                        "name": "farm_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Farm Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FarmSeason"
                                    },
                                    "title": "Response Get Farm Seasons Api V1 Farm  Farm Id  Seasons Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Farm": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "city": {
                        "type": "string",
                        "title": "City"
                    },
                    "state": {
                        "type": "string",
                        "title": "State"
                    },
                    "total_area": {
                        "type": "number",
                        "title": "Total Area"
                    },
                    "arable_area": {
                        "type": "number",
                        "title": "Arable Area"
                    },
                    "vegetation_area": {
                        "type": "number",
                        "title": "Vegetation Area"
                    },
                    "productor_id": {
                        "type": "integer",
                        "title": "Productor Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "city",
                    "state",
                    "total_area",
                    "arable_area",
                    "vegetation_area",
                    "productor_id",
                    "id"
                ],
                "title": "Farm"
            },
            "FarmCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "city": {
                        "type": "string",
                        "title": "City"
                    },
                    "state": {
                        "type": "string",
                        "title": "State"
                    },
                    "total_area": {
                        "type": "number",
                        "title": "Total Area"
                    },
                    "arable_area": {
                        "type": "number",
                        "title": "Arable Area"
                    },
                    "vegetation_area": {
                        "type": "number",
                        "title": "Vegetation Area"
                    },
                    "productor_id": {
                        "type": "integer",
                        "title": "Productor Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "city",
                    "state",
                    "total_area",
                    "arable_area",
                    "vegetation_area",
                    "productor_id"
                ],
                "title": "FarmCreate"
            },
            "FarmPlantation": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "seasons": {
                        "items": {
                            "$ref": "#/components/schemas/Season"
                        },
                        "type": "array",
                        "title": "Seasons"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "id",
                    "seasons"
                ],
                "title": "FarmPlantation"
            },
            "FarmPlantationSeason": {
                "properties": {
                    "plantation_id": {
                        "type": "integer",
                        "title": "Plantation Id"
                    },
                    "season_id": {
                        "type": "integer",
                        "title": "Season Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "farm_id": {
                        "type": "integer",
                        "title": "Farm Id"
                    }
                },
                "type": "object",
                "required": [
                    "plantation_id",
                    "season_id",
                    "id",
                    "farm_id"
                ],
                "title": "FarmPlantationSeason"
            },
            "FarmPlantationSeasonCreate": {
                "properties": {
                    "plantation_id": {
                        "type": "integer",
                        "title": "Plantation Id"
                    },
                    "season_id": {
                        "type": "integer",
                        "title": "Season Id"
                    }
                },
                "type": "object",
                "required": [
                    "plantation_id",
                    "season_id"
                ],
                "title": "FarmPlantationSeasonCreate"
            },
            "FarmSeason": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "year": {
                        "type": "integer",
                        "title": "Year"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "plantations": {
                        "items": {
                            "$ref": "#/components/schemas/Plantation"
                        },
                        "type": "array",
                        "title": "Plantations"
                    }
                },
                "type": "object",
                "required": [
                    "description",
                    "year",
                    "id",
                    "plantations"
                ],
                "title": "FarmSeason"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Plantation": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "description",
                    "id"
                ],
                "title": "Plantation"
            },
...
            "Season": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "year": {
                        "type": "integer",
                        "title": "Year"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "description",
                    "year",
                    "id"
                ],
                "title": "Season"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}